Refactor Teles Mediation
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C41109

+++ Mediation/Code/Mediation.Generic.Data.SQL/MediationRecordsDataManager.cs
@@ -128,7 +128,7 @@ namespace Mediation.Generic.Data.SQL
         {
             return new MediationRecord()
             {
-                EventId = (int)reader["EventId"],
+                EventId = (long)reader["EventId"],
                 SessionId = reader["SessionId"] as string,
                 EventTime = GetReaderValue<DateTime>(reader, "EventTime"),
                 EventStatus = (EventStatus)GetReaderValue<byte>(reader, "EventStatus"),
